fix: react-syntax-highlighter bundle size
Fixes an issue where all syntax highlighting languages were bundled into a single chunk, preventing effective code splitting and bloating the bundle.

### Changes
- Split package into default (light) and `/full` export paths
- Languages now load in separate chunks instead of one monolithic bundle
- Bundle size reduced from 811 KB to 307 KB (-62%)
### Breaking Changes
- No breaking changes, unless you were relying on languages that weren’t explicitly registered in the `syntax-highlighter.tsx` component. If you were, all you have to do is update the import path to `"@assistant-ui/react-syntax-highlighter/full”`, otherwise register langauges explicitly for a smaller bundle and faster first load

Fixes #1719
