+++ packages/react-syntax-highlighter/src/full.ts
@@ -1,6 +1,9 @@
 export {
+  makeSyntaxHighlighter,
   makeLightSyntaxHighlighter,
   makeLightAsyncSyntaxHighlighter,
+  makePrismSyntaxHighlighter,
   makePrismLightSyntaxHighlighter,
+  makePrismAsyncSyntaxHighlighter,
   makePrismAsyncLightSyntaxHighlighter,
-} from "./react-syntax-highlighter-light";
+} from "./react-syntax-highlighter-full";
