+++ packages/react-syntax-highlighter/src/react-syntax-highlighter-full.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import SyntaxHighlighter, {
+  Prism,
+  PrismAsync,
+  PrismAsyncLight,
+  PrismLight,
+  Light,
+  LightAsync,
+} from "react-syntax-highlighter";
+import { makeMakeSyntaxHighlighter } from "./make-syntax-highlighter";
+
+export const makeSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(SyntaxHighlighter);
+
+export const makePrismSyntaxHighlighter = makeMakeSyntaxHighlighter(Prism);
+
+export const makePrismAsyncSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(PrismAsync);
+
+export const makePrismAsyncLightSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(PrismAsyncLight);
+
+export const makePrismLightSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(PrismLight);
+
+export const makeLightSyntaxHighlighter = makeMakeSyntaxHighlighter(Light);
+
+export const makeLightAsyncSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(LightAsync);
