+++ packages/react-syntax-highlighter/src/react-syntax-highlighter-light.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import {
+  PrismAsyncLight,
+  PrismLight,
+  Light,
+  LightAsync,
+} from "react-syntax-highlighter";
+import { makeMakeSyntaxHighlighter } from "./make-syntax-highlighter";
+
+export const makePrismAsyncLightSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(PrismAsyncLight);
+
+export const makePrismLightSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(PrismLight);
+
+export const makeLightSyntaxHighlighter = makeMakeSyntaxHighlighter(Light);
+
+export const makeLightAsyncSyntaxHighlighter =
+  makeMakeSyntaxHighlighter(LightAsync);
